separate solutions into different files
